requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jwt==1.3.1
pycparser==2.21
PyJWT==2.4.0
pyotp==2.6.0
pytz==2022.2.1
simplejwt==2.0.1
sqlparse==0.4.2
typing==3.7.4.3
